@@ -1,48 +1,34 @@
-﻿using NUnit.Framework;
-using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using Nemesis.Essentials.Design;
+﻿using Nemesis.Essentials.Design;
 
-namespace Nemesis.Essentials.Tests
+namespace Nemesis.Essentials.Tests;
+
+[TestFixture(TestOf = typeof(EnumerableEqualityComparer<>))]
+public class EnumerableEqualityComparerTests
 {
-    [TestFixture(TestOf = typeof(EnumerableEqualityComparer<>))]
-    public class EnumerableEqualityComparerTests
+    private static IList<string> Empty() => new List<string>();
+    private static IList<string> From(params string[] elements) => new List<string>(elements);
+
+    private static IEnumerable<TCD> GetPositiveCases() => new[]
     {
-        private static IList<string> Empty() => new List<string>();
-        private static IList<string> From(params string[] elements) => new List<string>(elements);
-        [SuppressMessage("ReSharper", "PossibleMultipleEnumeration")]
-        private static string Format(IEnumerable<string> list) =>
-            list == null ? "NULL" :
-                (
-                    !list.Any() ? "∅" :
-                        string.Join("|", list.Select(FormatElement))
-                );
+        (null, null),
+        (Empty(), Empty()),
+        (Empty().AddFluent("1"),  Empty().AddFluent("1")),
+        (Empty().AddFluent("1", "2"),  Empty().AddFluent("1").AddFluent("2")),
+        (Empty().AddFluent("1", "2").AddFluent("3"),  Empty().AddFluent("1").AddFluent("2").AddFluent(3.ToString())),
+        (From("1", "2", "3"), From("1", "2", "3")),
+        (From("1", "2", "3", "4"), From("1", "2", "3").AddFluent("4")),
+    }.DuplicateWithElementReversal()
+        .Select((elem, i) => new TCD(elem.Item1, elem.Item2).SetName($"Pos_{i + 1:00}"));
 
-        private static string FormatElement(string element) =>
-            element == null ? "NULL_ELEMENT" : (element.Length == 0 ? "EMPTY_ELEMENT" : element);
+    [TestCaseSource(nameof(GetPositiveCases))]
+    public void Equals_Positive(IList<string> left, IList<string> right)
+    {
+        if (ReferenceEquals(left, right) && left != null && right != null)
+            Assert.Fail("Reference equality should be checked using trivial path");
 
-        private static IEnumerable<TestCaseData> GetPositiveCases() => new[]
+        var comparer = EnumerableEqualityComparer<string>.DefaultInstance;
+        Assert.Multiple(() =>
         {
-            (null, null),
-            (Empty(), Empty()),
-            (Empty().AddFluent("1"),  Empty().AddFluent("1")),
-            (Empty().AddFluent("1", "2"),  Empty().AddFluent("1").AddFluent("2")),
-            (Empty().AddFluent("1", "2").AddFluent("3"),  Empty().AddFluent("1").AddFluent("2").AddFluent(3.ToString())),
-            (From("1", "2", "3"), From("1", "2", "3")),
-            (From("1", "2", "3", "4"), From("1", "2", "3").AddFluent("4")),
-        }.DuplicateWithElementReversal()
-            .Select((elem, i) => new TestCaseData(elem).SetName($"{i + 1:00}. {Format(elem.Item1)} == {Format(elem.Item2)}"));
-
-        [TestCaseSource(nameof(GetPositiveCases))]
-        public void Equals_Positive((IList<string> left, IList<string> right) data)
-        {
-            var (left, right) = data;
-            if (ReferenceEquals(left, right) && left != null && right != null)
-                Assert.Fail("Reference equality should be checked using trivial path");
-
-            var comparer = EnumerableEqualityComparer<string>.DefaultInstance;
-
             Assert.That(left, Is.EqualTo(right).Using(comparer), "Equals assert");
 
             Assert.That(
@@ -51,48 +37,48 @@ namespace Nemesis.Essentials.Tests
                 ), "GetHashCode");
 
             Assert.That(comparer.Equals(left, right), Is.True, "Equals");
-        }
+        });
+    }
 
-
-        private static IEnumerable<TestCaseData> GetNegativeCases() => new[]
-            {
-                (Empty(), null),
-                (Empty().AddFluent("1"), null),
-                (From("1", "2", "3", "4"), From("1", "2", "3")),
-                (From("1", "2", "3", ""), From("1", "2", "3")),
-                (From("1", "2", "3", "4", "5"), From("1", "2", "3").AddFluent("4")),
-                (From("1", "2", "3", "4", "5", "6"), From("1", "2", "3").AddFluent("4")),
-            }.DuplicateWithElementReversal()
-            .Select((elem, i) => new TestCaseData(elem).SetName($"{i + 1:00}. {Format(elem.Item1)} != {Format(elem.Item2)}"));
-
-        [TestCaseSource(nameof(GetNegativeCases))]
-        public void Equals_Negative((IList<string> left, IList<string> right) data)
+    private static IEnumerable<TCD> GetNegativeCases() => new[]
         {
-            var (left, right) = data;
-            var comparer = EnumerableEqualityComparer<string>.DefaultInstance;
+            (Empty(), null),
+            (Empty().AddFluent("1"), null),
+            (From("1", "2", "3", "4"), From("1", "2", "3")),
+            (From("1", "2", "3", ""), From("1", "2", "3")),
+            (From("1", "2", "3", "4", "5"), From("1", "2", "3").AddFluent("4")),
+            (From("1", "2", "3", "4", "5", "6"), From("1", "2", "3").AddFluent("4")),
+        }.DuplicateWithElementReversal()
+        .Select((elem, i) => new TCD(elem.Item1, elem.Item2).SetName($"Neg_{i + 1:00}"));
 
+    [TestCaseSource(nameof(GetNegativeCases))]
+    public void Equals_Negative(IList<string> left, IList<string> right)
+    {
+        var comparer = EnumerableEqualityComparer<string>.DefaultInstance;
+        Assert.Multiple(() =>
+        {
             Assert.That(left, Is.Not.EqualTo(right).Using(comparer), "Equals assert");
 
             Assert.That(comparer.Equals(left, right), Is.False, "Equals");
-        }
+        });
+    }
+}
+
+file static class EnumerableEqualityComparerTestsHelper
+{
+    internal static IList<T> AddFluent<T>(this IList<T> list, params T[] elements)
+    {
+        foreach (var element in elements)
+            list.Add(element);
+        return list;
     }
 
-    internal static class EnumerableEqualityComparerTestsHelper
+    internal static IEnumerable<(T, T)> DuplicateWithElementReversal<T>(this IList<(T, T)> list)
     {
-        internal static IList<T> AddFluent<T>(this IList<T> list, params T[] elements)
-        {
-            foreach (var element in elements)
-                list.Add(element);
-            return list;
-        }
+        foreach (var t in list)
+            yield return t;
 
-        internal static IEnumerable<(T, T)> DuplicateWithElementReversal<T>(this IList<(T, T)> list)
-        {
-            foreach (var t in list)
-                yield return t;
-
-            for (int i = 0; i < list.Count; i++)
-                yield return (list[i].Item2, list[i].Item1);
-        }
+        for (int i = 0; i < list.Count; i++)
+            yield return (list[i].Item2, list[i].Item1);
     }
 }
