@@ -1,151 +1,139 @@
-﻿using System.Collections.Generic;
-using Nemesis.Essentials.Design;
-using NUnit.Framework;
+﻿using Nemesis.Essentials.Design;
 
-namespace Nemesis.Essentials.Tests
+namespace Nemesis.Essentials.Tests;
+
+[TestFixture(TestOf = typeof(EnumerableHelper))]
+public class EnumerableHelperTests
 {
-    [TestFixture(TestOf = typeof(EnumerableHelper))]
-    public class EnumerableHelperTests
+    [Test]
+    public void MultiDictionaryTests()
     {
-        [Test]
-        public void MultiDictionaryTests()
+        var multiDict = EnumerableHelper.CreateMultiDictionary<string, int>();
+        Assert.That(multiDict, Has.Count.EqualTo(0));
+
+
+        multiDict.Add("XXX", [1, 2, 3]);
+        Assert.Multiple(() =>
         {
-            var multiDict = EnumerableHelper.CreateMultiDictionary<string, int>();
-            Assert.That(multiDict, Has.Count.EqualTo(0));
-
-
-            multiDict.Add("XXX", new List<int> { 1, 2, 3 });
             Assert.That(multiDict.Keys, Has.Count.EqualTo(1));
             Assert.That(multiDict["XXX"], Has.Count.EqualTo(3));
-
-
-            multiDict.AddToMultiDictionary("XXX", 5);
+        });
+        multiDict.AddToMultiDictionary("XXX", 5);
+        Assert.Multiple(() =>
+        {
             Assert.That(multiDict.Keys, Has.Count.EqualTo(1));
             Assert.That(multiDict["XXX"], Has.Count.EqualTo(4));
-
-
-            multiDict.AddToMultiDictionary("XXX", 5, 6,7, 9 );
+        });
+        multiDict.AddToMultiDictionary("XXX", 5, 6, 7, 9);
+        Assert.Multiple(() =>
+        {
             Assert.That(multiDict.Keys, Has.Count.EqualTo(1));
             Assert.That(multiDict["XXX"], Has.Count.EqualTo(8));
-
-
-            multiDict.AddToMultiDictionary("YYY", 5, 6, 7, 9);
+        });
+        multiDict.AddToMultiDictionary("YYY", 5, 6, 7, 9);
+        Assert.Multiple(() =>
+        {
             Assert.That(multiDict.Keys, Has.Count.EqualTo(2));
             Assert.That(multiDict["YYY"], Has.Count.EqualTo(4));
-        }
+        });
+    }
 
-        [Test]
-        public void MultiSetDictionaryTests()
+    [Test]
+    public void MultiSetDictionaryTests()
+    {
+        var multiSetDict = EnumerableHelper.CreateMultiDictionaryUniqueElements<string, int>();
+        Assert.That(multiSetDict, Has.Count.EqualTo(0));
+
+
+        multiSetDict.Add("XXX", [1, 2, 3]);
+        Assert.Multiple(() =>
         {
-            var multiSetDict = EnumerableHelper.CreateMultiDictionaryUniqueElements<string, int>();
-            Assert.That(multiSetDict, Has.Count.EqualTo(0));
-
-
-            multiSetDict.Add("XXX", new HashSet<int>{ 1, 2, 3 });
             Assert.That(multiSetDict.Keys, Has.Count.EqualTo(1));
             Assert.That(multiSetDict["XXX"], Has.Count.EqualTo(3));
-
-
-            multiSetDict.AddToMultiDictionary("XXX", 5);
+        });
+        multiSetDict.AddToMultiDictionary("XXX", 5);
+        Assert.Multiple(() =>
+        {
             Assert.That(multiSetDict.Keys, Has.Count.EqualTo(1));
             Assert.That(multiSetDict["XXX"], Has.Count.EqualTo(4));
-
-
-            multiSetDict.AddToMultiDictionary("XXX", 5, 6, 7, 9, 7, 9, 5, 6);
+        });
+        multiSetDict.AddToMultiDictionary("XXX", 5, 6, 7, 9, 7, 9, 5, 6);
+        Assert.Multiple(() =>
+        {
             Assert.That(multiSetDict.Keys, Has.Count.EqualTo(1));
             Assert.That(multiSetDict["XXX"], Has.Count.EqualTo(7));
-
-
-            multiSetDict.AddToMultiDictionary("YYY", 5, 6, 7, 9);
+        });
+        multiSetDict.AddToMultiDictionary("YYY", 5, 6, 7, 9);
+        Assert.Multiple(() =>
+        {
             Assert.That(multiSetDict.Keys, Has.Count.EqualTo(2));
             Assert.That(multiSetDict["YYY"], Has.Count.EqualTo(4));
-        }
+        });
+    }
 
-        [Test]
-        public void ReverseIteratorTest()
+    [Test]
+    public void ReverseIteratorTest()
+    {
+        int[] array = [1, 2, 3];
+
+        using var it = array.AsReversedEnumerable().GetEnumerator();
+        var actual = it.ToArray(3);
+
+
+        Assert.Multiple(() =>
         {
-            int[] array = { 1, 2, 3 };
+            Assert.That(actual, Is.EqualTo(new[] { 3, 2, 1 }));
+            Assert.That(!it.MoveNext(), "end");
+        });
+    }
 
-            using (var it = array.AsReversedEnumerable().GetEnumerator())
-            {
-                it.MoveNext();
-                Assert.That(it.Current == 3, "3");
-                it.MoveNext();
-                Assert.That(it.Current == 2, "2");
-                it.MoveNext();
-                Assert.That(it.Current == 1, "1");
-                Assert.That(!it.MoveNext(), "end");
-            }
-        }
+    [Test]
+    public void CircularIteratorTest()
+    {
+        int[] array = [1, 2, 3];
 
-        [Test]
-        public void CircularIteratorTest()
+        using var it = array.AsCircularEnumerable().GetEnumerator();
+        var actual = it.ToArray(7);
+
+        Assert.Multiple(() =>
         {
-            int[] array = { 1, 2, 3 };
+            Assert.That(actual, Is.EqualTo(new[] { 1, 2, 3, 1, 2, 3, 1 }));
+            for (int i = 0; i < 10; i++)
+                Assert.That(it.MoveNext(), $"end {i}");
+        });
+    }
 
-            using (var it = array.AsCircularEnumerable().GetEnumerator())
-            {
-                it.MoveNext();
-                Assert.AreEqual(1, it.Current);
-                it.MoveNext();
-                Assert.AreEqual(2, it.Current);
-                it.MoveNext();
-                Assert.AreEqual(3, it.Current);
-                it.MoveNext();
-                Assert.AreEqual(1, it.Current);
-                it.MoveNext();
-                Assert.AreEqual(2, it.Current);
-            }
-        }
-
-        [Test]
-        public void SkippingIteratorTest()
+    [Test]
+    public void SkippingIteratorTest()
+    {
+        Dictionary<int, string> dict = new()
         {
-            Dictionary<int, string> dict = new Dictionary<int, string>
-            {
-                {1, "one"}, {2, "two"}, {3, "three"}, {4, "four"}, {5, "five"}, {6, "six"}, {7, "seven"}, {8, "eight"}, {9, "nine"}, {10, "ten"}
-            };
-            using (var it = dict.AsSkippingEnumerable(2).GetEnumerator())
-            {
-                it.MoveNext();
-                Assert.AreEqual(1, it.Current.Key);
-                it.MoveNext();
-                Assert.AreEqual(4, it.Current.Key);
-                it.MoveNext();
-                Assert.AreEqual(7, it.Current.Key);
-                it.MoveNext();
-                Assert.AreEqual(10, it.Current.Key);
+            {1, "one"}, {2, "two"}, {3, "three"}, {4, "four"}, {5, "five"}, {6, "six"}, {7, "seven"}, {8, "eight"}, {9, "nine"}, {10, "ten"}
+        };
+        using var it = dict.AsSkippingEnumerable(2).GetEnumerator();
+        var actual = it.ToArray(4).Select(kvp => kvp.Key);
 
-                Assert.That(!it.MoveNext(), "end");
-            }
-        }
 
+        Assert.Multiple(() =>
+        {
+            Assert.That(actual, Is.EqualTo(new[] { 1, 4, 7, 10 }));
+            Assert.That(!it.MoveNext(), "end");
+        });
     }
 }
-/*
 
-            Console.WriteLine("Constrained enumeration (2,5):");
-            foreach (KeyValuePair<int, string> pair in Enumerators.ConstrainedEnum(list, 2, 5))
-            {
-                Console.WriteLine("   " + pair.ToString());
-            }
-            Console.WriteLine("   (finished)\r\n");
+file static class Ext
+{
+    public static T[] ToArray<T>(this IEnumerator<T> enumerator, int length)
+    {
+        var array = new T[length];
 
-            Console.WriteLine("Stepped enumeration (3):");
-            foreach (int value in Enumerators.SteppedEnum(list.Keys, 3))
-            {
-                Console.WriteLine("   " + value.ToString());
-            }
-            Console.WriteLine("   (finished)\r\n");
+        for (int i = 0; i < length; i++)
+        {
+            enumerator.MoveNext();
+            array[i] = enumerator.Current;
+        }
 
-            Console.WriteLine("Combined Constrained(2,6)-Stepped(3)-Reverse-Circular enumeration:");
-            i = 0;
-            foreach (int value in Enumerators.ConstrainedEnum(Enumerators.SteppedEnum(Enumerators.CircularEnum(Enumerators.ReverseEnum(list.Keys)), 3), 2, 6))
-            {
-                Console.WriteLine("   " + value.ToString());
-                if (++i >= max * 2)
-                    break;   // stop circular enumerator, will be infinite if not
-            }
-            Console.WriteLine("   (finished)");
- 
- */
+        return array;
+    }
+}
